backend/requirements.txt
Flask==3.0.1
flask-cors==4.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
PyJWT==2.8.0
bcrypt==4.1.2
requests==2.31.0
python-socketio==5.10.0
